+++ src/test/java/ee/sk/smartid/rest/SmartIdConnectorSpy.java
@@ -26,10 +26,20 @@ package ee.sk.smartid.rest;
  * #L%
  */
 
-import ee.sk.smartid.exception.SessionNotFoundException;
-import ee.sk.smartid.rest.dao.*;
-
 import java.util.concurrent.TimeUnit;
+
+import javax.net.ssl.SSLContext;
+
+import ee.sk.smartid.exception.SessionNotFoundException;
+import ee.sk.smartid.rest.dao.AuthenticationSessionRequest;
+import ee.sk.smartid.rest.dao.AuthenticationSessionResponse;
+import ee.sk.smartid.rest.dao.CertificateChoiceResponse;
+import ee.sk.smartid.rest.dao.CertificateRequest;
+import ee.sk.smartid.rest.dao.NationalIdentity;
+import ee.sk.smartid.rest.dao.SemanticsIdentifier;
+import ee.sk.smartid.rest.dao.SessionStatus;
+import ee.sk.smartid.rest.dao.SignatureSessionRequest;
+import ee.sk.smartid.rest.dao.SignatureSessionResponse;
 
 public class SmartIdConnectorSpy implements SmartIdConnector {
 
@@ -113,4 +123,8 @@ public class SmartIdConnectorSpy implements SmartIdConnector {
 
   }
 
+  @Override
+  public void setSslContext(SSLContext sslContext) {
+
+  }
 }